+++ bsp_tool/lumps.py
@@ -1,4 +1,4 @@
-"""handles dynamically loading entries from lumps of all kinds"""
+"""classes for dynamically parsing lumps"""
 from __future__ import annotations
 
 import io
@@ -6,6 +6,13 @@ import lzma
 import struct
 from typing import Any, Dict, Union
 import warnings
+
+
+LumpHeader = Any
+# all: offset & length
+# ValveBsp / RespawnBsp: fourCC & version
+# external: filename & filesize
+Stream = Union[io.BufferedReader, io.BytesIO]
 
 
 def _remap_negative_index(index: int, length: int) -> int:
@@ -20,18 +27,9 @@ def _remap_negative_index(index: int, length: int) -> int:
 def _remap_slice(_slice: slice, length: int) -> slice:
     """simplify to positive start & stop within range(0, length)"""
     start, stop, step = _slice.start, _slice.stop, _slice.step
-    if start is None:
-        start = 0
-    elif start < 0:
-        start = max(length + start, 0)
-    if start > length:
-        start = length
-    if stop is None or stop > length:
-        stop = length
-    elif stop < 0:
-        stop = max(length + stop, 0)
-    if step is None:
-        step = 1
+    start = 0 if (start is None) else max(length + start, 0) if (start < 0) else length if (start > length) else start
+    stop = length if (stop is None or stop > length) else max(length + stop, 0) if (stop < 0) else stop
+    step = 1 if (step is None) else step
     return slice(start, stop, step)
 
 
@@ -44,69 +42,69 @@ def decompress_valve_LZMA(data: bytes) -> bytes:
     return decompressed_data[:true_size]  # trim any excess bytes
 
 
-def decompressed(file: io.BufferedReader, lump_header: Any) -> (io.BytesIO, Any):
+def decompressed(stream: Stream, lump_header: LumpHeader) -> (Stream, LumpHeader):
     """Takes a lump and decompresses it if nessecary. Also corrects lump_header offset & length"""
     if getattr(lump_header, "fourCC", 0) != 0:
-        # get lump bytes
-        if not hasattr(lump_header, "filename"):  # internal compressed lump
-            file.seek(lump_header.offset)
-            data = file.read(lump_header.length)
-        else:  # external compressed lump is unlikely, but possible
+        if not hasattr(lump_header, "filename"):
+            stream.seek(lump_header.offset)
+            data = stream.read(lump_header.length)
+        else:  # unlikely, but possible
             data = open(lump_header.filename, "rb").read()
-        file = io.BytesIO(decompress_valve_LZMA(data))
-        # modify lump header to point at the decompressed lump (fake file, data starts at offset=0)
+        stream = io.BytesIO(decompress_valve_LZMA(data))
         lump_header.offset = 0
         lump_header.length = lump_header.fourCC
-    return file, lump_header
+    return stream, lump_header
 
 
-def create_BspLump(file: io.BufferedReader, lump_header: Any, LumpClass: object = None) -> BspLump:
+def create_BspLump(stream: Stream, lump_header: LumpHeader, LumpClass: object = None) -> BspLump:
     if hasattr(lump_header, "fourCC"):
-        file, lump_header = decompressed(file, lump_header)
+        stream, lump_header = decompressed(stream, lump_header)
     if not hasattr(lump_header, "filename"):
-        if LumpClass is not None:
-            return BspLump(file, lump_header, LumpClass)
-        else:
-            return RawBspLump(file, lump_header)
+        return BspLump.from_header(stream, lump_header, LumpClass)
     else:
-        if LumpClass is not None:
-            return ExternalBspLump(lump_header, LumpClass)
-        else:
-            return ExternalRawBspLump(lump_header)
+        return ExternalBspLump.from_header(lump_header, LumpClass)
 
 
-def create_RawBspLump(file: io.BufferedReader, lump_header: Any) -> RawBspLump:
-    if not hasattr(lump_header, "filename"):
-        return RawBspLump(file, lump_header)
-    else:
-        return ExternalRawBspLump(lump_header)
-
-
-def create_BasicBspLump(file: io.BufferedReader, lump_header: Any, LumpClass: object) -> BasicBspLump:
+def create_RawBspLump(stream: Stream, lump_header: LumpHeader) -> RawBspLump:
     if hasattr(lump_header, "fourCC"):
-        file, lump_header = decompressed(file, lump_header)
+        stream, lump_header = decompressed(stream, lump_header)
     if not hasattr(lump_header, "filename"):
-        return BasicBspLump(file, lump_header, LumpClass)
+        return RawBspLump.from_header(stream, lump_header)
     else:
-        return ExternalBasicBspLump(lump_header, LumpClass)
+        return ExternalRawBspLump.from_header(lump_header)
+
+
+def create_BasicBspLump(stream: Stream, lump_header: LumpHeader, LumpClass: object) -> BasicBspLump:
+    if hasattr(lump_header, "fourCC"):
+        stream, lump_header = decompressed(stream, lump_header)
+    if not hasattr(lump_header, "filename"):
+        return BasicBspLump.from_header(stream, lump_header, LumpClass)
+    else:
+        return ExternalBasicBspLump.from_header(lump_header, LumpClass)
 
 
 class RawBspLump:
-    """Maps an open binary file to a list-like object"""
-    file: io.BufferedReader  # file opened in "rb" (read-bytes) mode
-    offset: int  # position in file where lump begins
+    """Maps an open binary stream to a bytearry-like object"""
+    # TODO: be more bytearray-like
+    stream: Stream
+    offset: int  # position in stream where lump begins
     _changes: Dict[int, bytes]
     # ^ {index: new_byte}
-    # NOTE: bytes objects don't support item assignment
-    # -- but we want to allow light hex editting with slice overrides
     _length: int  # number of indexable entries
 
-    def __init__(self, file: io.BufferedReader, lump_header: Any):
-        self.file = file
-        self.offset = lump_header.offset
+    def __init__(self):
         self._changes = dict()
-        # ^ {index: new_value}
-        self._length = lump_header.length
+        self._length = 0
+        self.offset = 0
+        self.stream = io.BytesIO(b"")
+
+    @classmethod
+    def from_header(cls, stream: Stream, lump_header: LumpHeader):
+        out = cls()
+        out._length = lump_header.length
+        out.offset = lump_header.offset
+        out.stream = stream
+        return out
 
     def __repr__(self):
         return f"<{self.__class__.__name__}; {len(self)} bytes at 0x{id(self):016X}>"
@@ -120,8 +118,8 @@ class RawBspLump:
             if index in self._changes:
                 return self._changes[index]
             else:
-                self.file.seek(self.offset + index)
-                return self.file.read(1)[0]  # return 1 0-255 integer, matching bytes behaviour
+                self.stream.seek(self.offset + index)
+                return self.stream.read(1)[0]  # return 1 0-255 integer, matching bytes behaviour
         elif isinstance(index, slice):
             _slice = _remap_slice(index, self._length)
             return bytes([self[i] for i in range(_slice.start, _slice.stop, _slice.step)])
@@ -166,25 +164,47 @@ class RawBspLump:
 
 
 class BspLump(RawBspLump):
-    """Dynamically reads LumpClasses from a binary file"""
-    file: io.BufferedReader  # file opened in "rb" (read-bytes) mode
-    offset: int  # position in file where lump begins
+    """Dynamically reads LumpClasses from a binary stream"""
+    stream: Stream
+    offset: int  # position in stream where lump begins
     LumpClass: object
-    _changes: Dict[int, Any]
+    _changes: Dict[int, object]
     # ^ {index: LumpClass(new_entry)}
     # NOTE: there are no checks to ensure changes are the correct type or size
     _entry_size: int  # sizeof(LumpClass)
     _length: int  # number of indexable entries
 
-    def __init__(self, file: io.BufferedReader, lump_header: Any, LumpClass: object):
-        self.file = file
-        self.offset = lump_header.offset
-        self._changes = dict()  # changes must be applied externally
-        self._entry_size = struct.calcsize(LumpClass._format)
-        if lump_header.length % self._entry_size != 0:
-            raise RuntimeError(f"LumpClass does not divide lump evenly! ({lump_header.length} / {self._entry_size})")
-        self._length = lump_header.length // self._entry_size
-        self.LumpClass = LumpClass
+    def __init__(self):
+        self._changes = dict()
+        self._entry_size = 0
+        self._length = 0
+        self.offset = 0
+        self.stream = io.BytesIO(b"")
+
+    @classmethod
+    def from_count(cls, stream: Stream, count: int, LumpClass: object):
+        """starts from cursor; stream.seek() before creating"""
+        out = cls()
+        out._entry_size = struct.sizeof(LumpClass._format)
+        out._length = count
+        out.offset = stream.tell()
+        out.stream = stream
+        byte_length = out._entry_size * out._length
+        if len(stream.read(byte_length)) != byte_length:
+            raise RuntimeError(f"{stream} is not long enough to hold {count} {LumpClass}")
+        return out
+
+    @classmethod
+    def from_header(cls, stream: Stream, lump_header: LumpHeader, LumpClass: object):
+        out = cls()
+        out.LumpClass = LumpClass
+        out._entry_size = struct.calcsize(LumpClass._format)
+        out._length = lump_header.length // out._entry_size
+        out.offset = lump_header.offset
+        out.stream = stream
+        if lump_header.length % out._entry_size != 0:
+            raise RuntimeError(f"LumpClass does not divide lump evenly! ({lump_header.length} / {out._entry_size})")
+        return out
 
     def __repr__(self):
         return f"<{self.__class__.__name__}({len(self)} {self.LumpClass.__name__}) at 0x{id(self):016X}>"
@@ -201,7 +221,7 @@ class BspLump(RawBspLump):
             raise TypeError(f"list indices must be integers or slices, not {type(index)}")
 
     def __getitem__(self, index: Union[int, slice]):
-        """Reads bytes from self.file & returns LumpClass(es)"""
+        """Reads bytes from self.stream & returns LumpClass(es)"""
         # read bytes -> struct.unpack tuples -> LumpClass
         # NOTE: BspLump[index] = LumpClass(entry)
         if isinstance(index, int):
@@ -209,8 +229,8 @@ class BspLump(RawBspLump):
             if index in self._changes:
                 return self._changes[index]
             else:
-                self.file.seek(self.offset + (index * self._entry_size))
-                _tuple = struct.unpack(self.LumpClass._format, self.file.read(self._entry_size))
+                self.stream.seek(self.offset + (index * self._entry_size))
+                _tuple = struct.unpack(self.LumpClass._format, self.stream.read(self._entry_size))
                 return self.LumpClass.from_tuple(_tuple)
         elif isinstance(index, slice):  # LAZY HACK
             _slice = _remap_slice(index, self._length)
@@ -250,9 +270,9 @@ class BspLump(RawBspLump):
 
 
 class BasicBspLump(BspLump):
-    """Dynamically reads LumpClasses from a binary file"""
-    file: io.BufferedReader  # file opened in "rb" (read-bytes) mode
-    offset: int  # position in file where lump begins
+    """Dynamically reads LumpClasses from a binary stream"""
+    stream: Stream
+    offset: int  # position in stream where lump begins
     LumpClass: object
     _changes: Dict[int, Any]
     # ^ {index: LumpClass(new_entry)}
@@ -261,13 +281,13 @@ class BasicBspLump(BspLump):
     _length: int  # number of indexable entries
 
     def __getitem__(self, index: Union[int, slice]):
-        """Reads bytes from self.file & returns LumpClass(es)"""
+        """Reads bytes from self.stream & returns LumpClass(es)"""
         # read bytes -> struct.unpack tuples -> LumpClass
         # NOTE: BspLump[index] = LumpClass(entry)
         if isinstance(index, int):
             index = _remap_negative_index(index, self._length)
-            self.file.seek(self.offset + (index * self._entry_size))
-            raw_entry = struct.unpack(self.LumpClass._format, self.file.read(self._entry_size))
+            self.stream.seek(self.offset + (index * self._entry_size))
+            raw_entry = struct.unpack(self.LumpClass._format, self.stream.read(self._entry_size))
             # NOTE: only the following line has changed
             return self.LumpClass(raw_entry[0])
         elif isinstance(index, slice):
@@ -281,54 +301,57 @@ class BasicBspLump(BspLump):
 
 
 class ExternalRawBspLump(RawBspLump):
-    """Maps an open binary file to a list-like object"""
-    file: io.BufferedReader  # file opened in "rb" (read-bytes) mode
-    offset: int  # position in file where lump begins
+    """Maps an open binary stream to a list-like object"""
+    stream: Stream
+    offset: int  # position in stream where lump begins
     _changes: Dict[int, bytes]
     # ^ {index: new_byte}
     _length: int  # number of indexable entries
     # -- should also override any returned entries with _changes
 
-    def __init__(self, lump_header: Any):
-        self.file = open(lump_header.filename, "rb")
-        self.offset = 0
-        self._changes = dict()  # changes must be applied externally
-        self._length = lump_header.filesize
+    @classmethod
+    def from_header(cls, lump_header: LumpHeader):
+        out = cls()
+        out._length = lump_header.filesize
+        out.offset = 0
+        out.stream = open(lump_header.filename, "rb")
+        return out
 
 
 class ExternalBspLump(BspLump):
-    """Dynamically reads LumpClasses from a binary file"""
-    # NOTE: this class does not handle compressed lumps
-    file: io.BufferedReader  # file opened in "rb" (read-bytes) mode
-    offset: int  # position in file where lump begins
+    """Dynamically reads LumpClasses from a binary stream"""
+    stream: Stream
+    offset: int  # position in stream where lump begins
     LumpClass: object
-    _changes: Dict[int, Any]
+    _changes: Dict[int, object]
     # ^ {index: LumpClass(new_entry)}
     # NOTE: there are no checks to ensure changes are the correct type or size
     _entry_size: int  # sizeof(LumpClass)
     _length: int  # number of indexable entries
 
-    def __init__(self, lump_header: Any, LumpClass: object):
-        super(ExternalBspLump, self).__init__(None, lump_header, LumpClass)
-        self.file = open(lump_header.filename, "rb")
-        self.offset = 0  # NOTE: 16 if ValveBsp
-        self._changes = dict()  # changes must be applied externally
+    @classmethod
+    def from_header(cls, lump_header: LumpHeader, LumpClass: object):
+        out = super().from_header(None, lump_header, LumpClass)
+        out.offset = 0
+        out.stream = open(lump_header.filename, "rb")
+        return out
 
 
 class ExternalBasicBspLump(BasicBspLump):
-    """Dynamically reads LumpClasses from a binary file"""
-    # NOTE: this class does not handle compressed lumps
-    file: io.BufferedReader  # file opened in "rb" (read-bytes) mode
-    offset: int  # position in file where lump begins
-    _changes: Dict[int, Any]
+    """Dynamically reads LumpClasses from a binary stream"""
+    stream: Stream
+    offset: int  # position in stream where lump begins
+    LumpClass: object
+    _changes: Dict[int, object]
     # ^ {index: LumpClass(new_entry)}
     _length: int  # number of indexable entries
 
-    def __init__(self, lump_header: Any, LumpClass: object):
-        super(ExternalBasicBspLump, self).__init__(None, lump_header, LumpClass)
-        self.file = open(lump_header.filename, "rb")
-        self.offset = 0
-        self._changes = dict()  # changes must be applied externally
+    @classmethod
+    def from_header(cls, lump_header: LumpHeader, LumpClass: object):
+        out = super().__init__(None, lump_header, LumpClass)
+        out.offset = 0
+        out.stream = open(lump_header.filename, "rb")
+        return out
 
 
 class GameLump:
@@ -347,22 +370,22 @@ class GameLump:
     # -- dprp: Detail Props (procedural grass on displacements)
     # -- sprp: Static Props
 
-    def __init__(self, file: io.BufferedReader, lump_header: Any, endianness: str,
+    def __init__(self, stream: Stream, lump_header: LumpHeader, endianness: str,
                  LumpClasses: Dict[str, object], GameLumpHeaderClass: object):
-        self.endianness = endianness
         self.GameLumpHeaderClass = GameLumpHeaderClass
+        self.endianness = endianness
         self.loading_errors = dict()
         lump_offset = 0
         if not hasattr(lump_header, "filename"):
             lump_offset = lump_header.offset
-            file.seek(lump_offset)
+            stream.seek(lump_offset)
         else:
             self.is_external = True
-        game_lumps_count = int.from_bytes(file.read(4), self.endianness)
+        game_lumps_count = int.from_bytes(stream.read(4), self.endianness)
         self.headers = dict()
         # {"child_name": child_header}
         for i in range(game_lumps_count):
-            child_header = GameLumpHeaderClass.from_stream(file)
+            child_header = GameLumpHeaderClass.from_stream(stream)
             if self.is_external:
                 child_header.offset = child_header.offset - lump_header.offset
             child_name = child_header.id.decode("ascii")
@@ -375,11 +398,11 @@ class GameLump:
         for child_name, child_header in self.headers.items():
             child_LumpClass = LumpClasses.get(child_name, dict()).get(child_header.version, None)
             if child_LumpClass is None:
-                setattr(self, child_name, create_RawBspLump(file, child_header))
+                setattr(self, child_name, create_RawBspLump(stream, child_header))
             else:
-                file.seek(child_header.offset)
+                stream.seek(child_header.offset)
                 try:
-                    child_lump_bytes = file.read(child_header.length)
+                    child_lump_bytes = stream.read(child_header.length)
                     compressed = child_lump_bytes[:4] == b"LZMA"
                     if compressed and child_header.flags & 1 != 1:
                         warnings.warn(UserWarning(f"{child_name} game lump is compressed but the flag is unset (Xbox360?)"))
@@ -394,17 +417,17 @@ class GameLump:
                         child_lump = None
                 except Exception as exc:
                     self.loading_errors[child_name] = exc
-                    child_lump = create_RawBspLump(file, child_header)
+                    child_lump = create_RawBspLump(stream, child_header)
                     # NOTE: RawBspLumps are not decompressed
                 if child_lump is not None:  # discarding compressed empty lumps
                     setattr(self, child_name, child_lump)
                 # TODO: test how empty lumps are written back
                 # -- can the branch that produced this wierd case function without this lump?
         if self.is_external:
-            file.close()
+            stream.close()
 
     def as_bytes(self, lump_offset=0):
-        """lump_offset makes headers relative to the file"""
+        """lump_offset makes headers relative to the stream"""
         # NOTE: ValveBsp .lmp external lumps have a 16 byte header
         # NOTE: RespawnBsp .bsp_lump offsets are relative to the internal .bsp GAME_LUMP.offset
         # NOTE: Xbox360 child lumps will not be recompressed
@@ -444,7 +467,7 @@ class DarkMessiahSPGameLump:
     # ^ {"child_lump": Error}
     unknown: int
 
-    def __init__(self, file: io.BufferedReader, lump_header: Any, endianness: str,
+    def __init__(self, stream: io.BufferedReader, lump_header: Any, endianness: str,
                  LumpClasses: Dict[str, object], GameLumpHeaderClass: object):
         self.endianness = endianness
         self.GameLumpHeaderClass = GameLumpHeaderClass
@@ -452,15 +475,15 @@ class DarkMessiahSPGameLump:
         lump_offset = 0
         if not hasattr(lump_header, "filename"):
             lump_offset = lump_header.offset
-            file.seek(lump_offset)
+            stream.seek(lump_offset)
         else:
             self.is_external = True
-        game_lumps_count = int.from_bytes(file.read(4), self.endianness)
-        self.unknown = int.from_bytes(file.read(4), self.endianness)
+        game_lumps_count = int.from_bytes(stream.read(4), self.endianness)
+        self.unknown = int.from_bytes(stream.read(4), self.endianness)
         self.headers = dict()
         # {"child_name": child_header}
         for i in range(game_lumps_count):
-            child_header = GameLumpHeaderClass.from_stream(file)
+            child_header = GameLumpHeaderClass.from_stream(stream)
             if self.is_external:
                 child_header.offset = child_header.offset - lump_header.offset
             child_name = child_header.id.decode("ascii")
@@ -473,11 +496,11 @@ class DarkMessiahSPGameLump:
         for child_name, child_header in self.headers.items():
             child_LumpClass = LumpClasses.get(child_name, dict()).get(child_header.version, None)
             if child_LumpClass is None:
-                setattr(self, child_name, create_RawBspLump(file, child_header))
+                setattr(self, child_name, create_RawBspLump(stream, child_header))
             else:
-                file.seek(child_header.offset)
+                stream.seek(child_header.offset)
                 try:
-                    child_lump_bytes = file.read(child_header.length)
+                    child_lump_bytes = stream.read(child_header.length)
                     # check if GameLump child is LZMA compressed (Xbox360)
                     # -- GameLumpHeader.flags does not appear to inidicate compression
                     if child_lump_bytes[:4] == b"LZMA":
@@ -485,14 +508,13 @@ class DarkMessiahSPGameLump:
                     child_lump = child_LumpClass.from_bytes(child_lump_bytes)
                 except Exception as exc:
                     self.loading_errors[child_name] = exc
-                    child_lump = create_RawBspLump(file, child_header)
-                    # NOTE: RawBspLumps are not decompressed
+                    child_lump = create_RawBspLump(stream, child_header)
                 setattr(self, child_name, child_lump)
         if self.is_external:
-            file.close()
+            stream.close()
 
     def as_bytes(self, lump_offset=0) -> bytes:
-        """lump_offset makes headers relative to the file"""
+        """lump_offset makes headers relative to the stream"""
         # NOTE: ValveBsp .lmp external lumps have a 16 byte header
         # NOTE: RespawnBsp .bsp_lump offsets are relative to the internal .bsp GAME_LUMP.offset
         # NOTE: Xbox360 child lumps will not be recompressed